# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/dishes/{id}:
    get:
      tags:
      - dish-rest-controller
      summary: Get a dish by its id
      operationId: findDishById
      parameters:
      - name: id
        in: path
        description: id of dish to be searched
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found the dish
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "400":
          description: Invalid id supplied
        "404":
          description: Dish not found
    put:
      tags:
      - dish-rest-controller
      summary: Update a dish not including the image (only admin can do that)
      operationId: replaceDish
      parameters:
      - name: id
        in: path
        description: id of dish to be updated
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dish'
        required: true
      responses:
        "400":
          description: Invalid form of introducing the data for updating the dish
        "200":
          description: The dish has been updated correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "404":
          description: Dish couldn't be updated
        "403":
          description: Forbidden. You have to be an admin to do this
    delete:
      tags:
      - dish-rest-controller
      summary: Delete a dish by its id (only admin can do that)
      operationId: removeDish
      parameters:
      - name: id
        in: path
        description: id of the dish to be deleted
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid id supplied
        "200":
          description: The dish has been deleted correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "403":
          description: Forbidden. You have to be an admin to do this
        "404":
          description: Dish couldn't be deleted
...
  /api/dishes/{id}/image:
    get:
      tags:
      - dish-rest-controller
      summary: Download the image of a dish
      operationId: downloadImage
      parameters:
      - name: id
        in: path
        description: id of dish
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Dish image couldn't be downloaded
        "200":
          description: The dish image has been downloaded correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
    post:
      tags:
      - dish-rest-controller
      summary: Upload the image of a dish (only admin can do that)
      operationId: uploadDishImage
      parameters:
      - name: id
        in: path
        description: id of dish
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  description: image of the dish
                  format: binary
      responses:
        "400":
          description: Invalid form of introducing the image of the dish
        "404":
          description: Dish image couldn't be uploaded
        "403":
          description: Forbidden. You have to be an admin to do this
        "200":
          description: The dish image has been uploaded correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
    delete:
      tags:
      - dish-rest-controller
      summary: Delete the image of a dish (only admin can do that)
      operationId: deleteImage
      parameters:
      - name: id
        in: path
        description: id of dish
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid id supplied
        "404":
          description: Dish image couldn't be deleted
        "204":
          description: No content. The dish image has been deleted correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "403":
          description: Forbidden. You have to be an admin to do this
  /api/dishes/:
    get:
      tags:
      - dish-rest-controller
      summary: Get all dishes
      operationId: findDishes
      responses:
        "404":
          description: Dishes not found
        "200":
          description: Found all dishes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
    post:
      tags:
      - dish-rest-controller
      summary: Upload a new dish without an image (only admin can do that)
      operationId: newDishProcess
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dish'
        required: true
      responses:
        "201":
          description: The dish has been uploaded correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "404":
          description: Dish couldn't be created
        "400":
          description: Invalid form of introducing the data for the new dish
        "403":
          description: Forbidden. You have to be an admin to do this
...
  /api/dishes/category:
    get:
      tags:
      - dish-rest-controller
      summary: Get all dishes from a category
      operationId: findDishesByTipe
      parameters:
      - name: category
        in: query
        description: "category of dish to be searched (desayuno, comida or cena)"
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Dishes from category not found
        "200":
          description: Found the dishes from a category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
        "400":
          description: Invalid category supplied
components:
  schemas:
    Dish:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        dishPrice:
          type: number
          format: float
        category:
          type: string
        nbuy:
          type: integer
          format: int32
        image:
          type: boolean
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        nameAllergen:
          type: string
        allergen:
          type: boolean
...